Split claims and reclaims to different azure function instances
