+++ src/HE.FMS.Middleware.Claims/Functions/CreateClaimHttpTrigger.cs
@@ -1,13 +1,16 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
 using HE.FMS.Middleware.Common.Serialization;
 using HE.FMS.Middleware.Contract.Claims;
-using HE.FMS.Middleware.Functions.Base;
-using HE.FMS.Middleware.Middlewares;
 using HE.FMS.Middleware.Providers.CosmosDb;
 using HE.FMS.Middleware.Providers.ServiceBus;
+using HE.FMS.Middleware.Shared.Base;
+using HE.FMS.Middleware.Shared.Middlewares;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 
-namespace HE.FMS.Middleware.Functions.HttpTriggers.Claims;
+namespace HE.FMS.Middleware.Claims.Functions;
 public class CreateClaimHttpTrigger : ClaimBase<CreateClaimRequest>
 {
     public CreateClaimHttpTrigger(
