+++ src/HE.FMS.Middleware.Claims/Properties/launchSettings.json
@@ -1,0 +1,9 @@
+{
+  "profiles": {
+    "HE.FMS.Middleware.Claims": {
+      "commandName": "Project",
+      "commandLineArgs": "--port 7180",
+      "launchBrowser": false
+    }
+  }
+}
