+++ src/HE.FMS.Middleware.Reclaims/Functions/CreateReclaimHttpTrigger.cs
@@ -1,13 +1,16 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
 using HE.FMS.Middleware.Common.Serialization;
 using HE.FMS.Middleware.Contract.Reclaims;
-using HE.FMS.Middleware.Functions.Base;
-using HE.FMS.Middleware.Middlewares;
 using HE.FMS.Middleware.Providers.CosmosDb;
 using HE.FMS.Middleware.Providers.ServiceBus;
+using HE.FMS.Middleware.Shared.Base;
+using HE.FMS.Middleware.Shared.Middlewares;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 
-namespace HE.FMS.Middleware.Functions.HttpTriggers.Reclaims;
+namespace HE.FMS.Middleware.Reclaims.Functions;
 public class CreateReclaimHttpTrigger : ClaimBase<CreateReclaimRequest>
 {
     public CreateReclaimHttpTrigger(
