+++ src/HE.FMS.Middleware.Reclaims/Properties/launchSettings.json
@@ -1,0 +1,12 @@
+{
+  "profiles": {
+    "HE.FMS.Middleware.Reclaims": {
+      "commandName": "Project",
+      "commandLineArgs": "--port 7242",
+      "launchBrowser": false,
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development"
+      }
+    }
+  }
+}
