+++ src/HE.FMS.Middleware.Shared/Middlewares/ExceptionHandlingForServiceBusTriggersMiddleware.cs
@@ -1,16 +1,19 @@
+using System;
 using System.Diagnostics.CodeAnalysis;
+using System.Linq;
+using System.Threading.Tasks;
 using HE.FMS.Middleware.Common;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Middleware;
 using Microsoft.Extensions.Logging;
 
-namespace HE.FMS.Middleware.Middlewares;
+namespace HE.FMS.Middleware.Shared.Middlewares;
 
-internal sealed class ExceptionHandlingWithoutResponseMiddleware : IFunctionsWorkerMiddleware
+internal sealed class ExceptionHandlingForServiceBusTriggersMiddleware : IFunctionsWorkerMiddleware
 {
-    private readonly ILogger<ExceptionHandlingWithoutResponseMiddleware> _logger;
+    private readonly ILogger<ExceptionHandlingForServiceBusTriggersMiddleware> _logger;
 
-    public ExceptionHandlingWithoutResponseMiddleware(ILogger<ExceptionHandlingWithoutResponseMiddleware> logger)
+    public ExceptionHandlingForServiceBusTriggersMiddleware(ILogger<ExceptionHandlingForServiceBusTriggersMiddleware> logger)
     {
         _logger = logger ?? throw new ArgumentNullException(nameof(logger));
     }
@@ -20,9 +23,7 @@ internal sealed class ExceptionHandlingWithoutResponseMiddleware : IFunctionsWor
     {
         if (context.FunctionDefinition.InputBindings.Values
              .Any(binding =>
-                    binding.Type.EndsWith(Constants.FunctionsTriggers.TimeTrigger, StringComparison.OrdinalIgnoreCase) ||
-                    binding.Type.EndsWith(Constants.FunctionsTriggers.OrchestrationTrigger, StringComparison.OrdinalIgnoreCase) ||
-                    binding.Type.EndsWith(Constants.FunctionsTriggers.ActivityTrigger, StringComparison.OrdinalIgnoreCase)))
+                    binding.Type.EndsWith(Constants.FunctionsTriggers.ServiceBusTrigger, StringComparison.OrdinalIgnoreCase)))
         {
             try
             {
@@ -31,6 +32,7 @@ internal sealed class ExceptionHandlingWithoutResponseMiddleware : IFunctionsWor
             catch (Exception ex)
             {
                 _logger.LogError(ex, ex.Message);
+                throw;
             }
         }
         else
