+++ src/HE.FMS.Middleware.Shared/Middlewares/RequestLoggingMiddleware.cs
@@ -1,8 +1,9 @@
-﻿using Microsoft.Azure.Functions.Worker;
+using System.Threading.Tasks;
+using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Middleware;
 using Microsoft.Extensions.Logging;
 
-namespace HE.FMS.Middleware.Middlewares;
+namespace HE.FMS.Middleware.Shared.Middlewares;
 
 internal sealed class RequestLoggingMiddleware : IFunctionsWorkerMiddleware
 {
