+++ src/HE.FMS.Middleware.sln
@@ -25,11 +25,17 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{A5EDFBA3
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HE.FMS.Middleware.Common.Tests", "HE.FMS.Middleware.Common.Tests\HE.FMS.Middleware.Common.Tests.csproj", "{10E380B9-769D-4A8F-B726-EBFD84095D66}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HE.FMS.Middleware.BusinessLogic.Tests", "HE.FMS.Middleware.BusinessLogic.Tests\HE.FMS.Middleware.BusinessLogic.Tests.csproj", "{6D8477EC-6324-49FD-8835-65D4B821E2FA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HE.FMS.Middleware.BusinessLogic.Tests", "HE.FMS.Middleware.BusinessLogic.Tests\HE.FMS.Middleware.BusinessLogic.Tests.csproj", "{6D8477EC-6324-49FD-8835-65D4B821E2FA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HE.FMS.Middleware.Providers.Tests", "HE.FMS.Middleware.Providers.Tests\HE.FMS.Middleware.Providers.Tests.csproj", "{A2947302-1BDB-46C2-835E-C93CD97F9A5C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HE.FMS.Middleware.Providers.Tests", "HE.FMS.Middleware.Providers.Tests\HE.FMS.Middleware.Providers.Tests.csproj", "{A2947302-1BDB-46C2-835E-C93CD97F9A5C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HE.FMS.Middleware.Tests", "HE.FMS.Middleware.Tests\HE.FMS.Middleware.Tests.csproj", "{19C136C7-04B2-406B-8756-AB7815BFCE60}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HE.FMS.Middleware.Tests", "HE.FMS.Middleware.Tests\HE.FMS.Middleware.Tests.csproj", "{19C136C7-04B2-406B-8756-AB7815BFCE60}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HE.FMS.Middleware.Claims", "HE.FMS.Middleware.Claims\HE.FMS.Middleware.Claims.csproj", "{4588FA1F-1D6A-441A-9A46-FF229FC56559}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HE.FMS.Middleware.Reclaims", "HE.FMS.Middleware.Reclaims\HE.FMS.Middleware.Reclaims.csproj", "{A38414A7-939A-4473-BD63-98933DA29925}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HE.FMS.Middleware.Shared", "HE.FMS.Middleware.Shared\HE.FMS.Middleware.Shared.csproj", "{F79BD2EB-B8C9-446C-97F1-2465B8726026}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -73,6 +79,18 @@ Global
 		{19C136C7-04B2-406B-8756-AB7815BFCE60}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{19C136C7-04B2-406B-8756-AB7815BFCE60}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{19C136C7-04B2-406B-8756-AB7815BFCE60}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4588FA1F-1D6A-441A-9A46-FF229FC56559}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4588FA1F-1D6A-441A-9A46-FF229FC56559}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4588FA1F-1D6A-441A-9A46-FF229FC56559}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4588FA1F-1D6A-441A-9A46-FF229FC56559}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A38414A7-939A-4473-BD63-98933DA29925}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A38414A7-939A-4473-BD63-98933DA29925}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A38414A7-939A-4473-BD63-98933DA29925}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A38414A7-939A-4473-BD63-98933DA29925}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F79BD2EB-B8C9-446C-97F1-2465B8726026}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F79BD2EB-B8C9-446C-97F1-2465B8726026}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F79BD2EB-B8C9-446C-97F1-2465B8726026}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F79BD2EB-B8C9-446C-97F1-2465B8726026}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
