+++ src/HE.FMS.Middleware/Functions/HttpTriggers/Mambu/MambuHealthCheckHttpTrigger.cs
@@ -1,4 +1,6 @@
-﻿using System.Net;
+using System.Net;
+using System.Threading;
+using System.Threading.Tasks;
 using HE.FMS.Middleware.BusinessLogic.Mambu;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
