+++ src/HE.FMS.Middleware/Functions/TimeTriggers/MambuRotateApiKeyTimeTrigger.cs
@@ -1,4 +1,6 @@
-﻿using HE.FMS.Middleware.BusinessLogic.Mambu;
+using System.Threading;
+using System.Threading.Tasks;
+using HE.FMS.Middleware.BusinessLogic.Mambu;
 using Microsoft.Azure.Functions.Worker;
 
 namespace HE.FMS.Middleware.Functions.TimeTriggers;
