+++ src/HE.FMS.Middleware/local.settings.json
@@ -1,5 +1,5 @@
 {
-    "IsEncrypted": false,
+  "IsEncrypted": false,
   "Values": {
     "AzureWebJobsStorage": "UseDevelopmentStorage=true",
     "FUNCTIONS_WORKER_RUNTIME": "dotnet-isolated",
@@ -25,8 +25,6 @@
     "Grants:BranchId": "{mambuEnvironment}",
     "Grants:OpenGrantAccount:TopicName": "he-ahp-CreateGrantAccount-local",
     "Grants:OpenGrantAccount:SubscriptionName": "subscription-GrantCreate-local",
-    "Claims:Create:TopicName": "he-ahp-CreateClaim-local",
-    "Reclaims:Create:TopicName": "he-ahp-CreateReclaim-local",
     "AzureWebJobs.OpenNewGrantAccountHttpTrigger.Disabled": true,
     "AzureWebJobs.OpenNewGrantAccountServiceBusTrigger.Disabled": true,
     "AzureWebJobs.PushToCrmServiceBusTrigger.Disabled": true,
